# docs/swagger.yaml
basePath: /
definitions:
...
  web.PokinOpdCreateRequest:
    properties:
      kode_opd:
        type: string
      nama_opd:
        type: string
      tahun:
        type: integer
    type: object
  web.PokinOpdResponse:
    properties:
      id:
        type: integer
      kode_opd:
        type: string
      nama_opd:
        type: string
      tahun:
        type: integer
    type: object
...
  web.PokinOpdUpdateRequest:
    properties:
      id:
        type: integer
      kode_opd:
        type: string
      nama_opd:
        type: string
      tahun:
        type: integer
    required:
    - id
    type: object
  web.TargetPokinOpdCreateRequest:
    properties:
      indikator_pokin_opd_id:
        type: integer
      satuan:
        type: string
      target:
        type: integer
    type: object
  web.TargetPokinOpdResponse:
    properties:
      id:
        type: integer
      satuan:
        type: string
      target:
        type: integer
    type: object
...
  web.TargetPokinOpdUpdateRequest:
    properties:
      id:
        type: integer
      satuan:
        type: string
      target:
        type: integer
    required:
    - id
    type: object
...
  web.WebResponse:
    description: Web Response
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
host: ${PROD_HOSTNAME}
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API For Alur Kerja Services
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Alur Kerja Service API
  version: "1.0"
paths:
...
  /pokin-opds:
    get:
      consumes:
      - application/json
      description: Get list of all Pokin Opd
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.PokinOpdResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: List All Pokin Opd
      tags:
      - Pokin Opd
    post:
      consumes:
      - application/json
      description: Create new Pokin Opd
      parameters:
      - description: Pokin Opd Create Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.PokinOpdCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.PokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Create Pokin Opd
      tags:
      - Pokin Opd
  /pokin-opds/{id}:
    delete:
      consumes:
      - application/json
      description: Delete existing Pokin Opd by ID
      parameters:
      - description: Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.PokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Delete Pokin Opd
      tags:
      - Pokin Opd
    get:
      consumes:
      - application/json
      description: Get Pokin Opd detail by ID
      parameters:
      - description: Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.PokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Pokin Opd by ID
      tags:
      - Pokin Opd
    put:
      consumes:
      - application/json
      description: Update existing Pokin Opd by ID
      parameters:
      - description: Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      - description: Pokin Opd Update Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.PokinOpdUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.PokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Update Pokin Opd
      tags:
      - Pokin Opd
...
  /target-pokin-opds:
    get:
      consumes:
      - application/json
      description: Get list of all Target Pokin Opd
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.TargetPokinOpdResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: List All Target Pokin Opd
      tags:
      - Target Pokin Opd
    post:
      consumes:
      - application/json
      description: Create new Target Pokin Opd
      parameters:
      - description: Target Pokin Opd Create Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.TargetPokinOpdCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TargetPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Create Target Pokin Opd
      tags:
      - Target Pokin Opd
  /target-pokin-opds/{id}:
    delete:
      consumes:
      - application/json
      description: Delete existing Target Pokin Opd by ID
      parameters:
      - description: Target Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Delete Target Pokin Opd
      tags:
      - Target Pokin Opd
    get:
      consumes:
      - application/json
      description: Get Target Pokin Opd detail by ID
      parameters:
      - description: Target Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TargetPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Target Pokin Opd by ID
      tags:
      - Target Pokin Opd
    put:
      consumes:
      - application/json
      description: Update existing Target Pokin Opd by ID
      parameters:
      - description: Target Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target Pokin Opd Update Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.TargetPokinOpdUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TargetPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Update Target Pokin Opd
      tags:
      - Target Pokin Opd
...
schemes:
- http
- https
swagger: "2.0"